restructured extensions in profile module. Created separated files under profile/data/model/extensions/
