+++ app/src/main/java/com/babylon/wallet/android/presentation/account/settings/devsettings/DevSettingsViewModel.kt
@@ -23,8 +23,8 @@ import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import rdx.works.core.UUIDGenerator
+import rdx.works.profile.data.model.extensions.hasAuthSigning
 import rdx.works.profile.data.model.pernetwork.Network
-import rdx.works.profile.data.utils.hasAuthSigning
 import rdx.works.profile.domain.GetProfileUseCase
 import rdx.works.profile.domain.account.AddAuthSigningFactorInstanceUseCase
 import rdx.works.profile.domain.accountsOnCurrentNetwork
