+++ app/src/main/java/com/babylon/wallet/android/presentation/transfer/prepare/PrepareManifestDelegate.kt
@@ -16,6 +16,7 @@ import kotlinx.coroutines.flow.update
 import rdx.works.core.ret.BabylonManifestBuilder
 import rdx.works.core.ret.buildSafely
 import rdx.works.core.toRETDecimal
+import rdx.works.profile.data.model.extensions.hasAnyDenyDepositRule
 import rdx.works.profile.data.model.factorsources.FactorSourceKind
 import rdx.works.profile.data.model.pernetwork.Network
 import rdx.works.profile.data.repository.MnemonicRepository
@@ -145,7 +146,7 @@ class PrepareManifestDelegate(
         } ?: false
 
         if (targetAccount is TargetAccount.Owned && isAccountAbleToSign) {
-            if (targetAccount.hasAnyDenyDepositRule(forSpecificAssetAddress = spendingAsset.address)) {
+            if (targetAccount.account.hasAnyDenyDepositRule(forSpecificAssetAddress = spendingAsset.address)) {
                 // if for example account has deny all we don't want to prevent transfer between our OWN accounts
                 // therefore ask user to sign
                 accountDeposit(
