+++ profile/src/main/java/rdx/works/profile/data/model/apppreferences/Security.kt
@@ -2,7 +2,6 @@ package rdx.works.profile.data.model.apppreferences
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import rdx.works.profile.data.model.Profile
 
 @Serializable
 data class Security(
@@ -23,19 +22,3 @@ data class Security(
         )
     }
 }
-
-fun Profile.updateDeveloperMode(isEnabled: Boolean): Profile = copy(
-    appPreferences = appPreferences.copy(
-        security = appPreferences.security.copy(
-            isDeveloperModeEnabled = isEnabled
-        )
-    )
-)
-
-fun Profile.updateCloudSyncEnabled(isEnabled: Boolean) = copy(
-    appPreferences = appPreferences.copy(
-        security = appPreferences.security.copy(
-            isCloudProfileSyncEnabled = isEnabled
-        )
-    )
-)
