+++ profile/src/main/java/rdx/works/profile/domain/ChangeEntityVisibilityUseCase.kt
@@ -3,11 +3,11 @@ package rdx.works.profile.domain
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
+import rdx.works.profile.data.model.extensions.hideAccount
+import rdx.works.profile.data.model.extensions.hidePersona
+import rdx.works.profile.data.model.extensions.unhideAllEntities
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
-import rdx.works.profile.data.utils.hideAccount
-import rdx.works.profile.data.utils.hidePersona
-import rdx.works.profile.data.utils.unhideAllEntities
 import rdx.works.profile.di.coroutines.DefaultDispatcher
 import javax.inject.Inject
