+++ profile/src/main/java/rdx/works/profile/domain/account/RenameAccountDisplayNameUseCase.kt
@@ -3,10 +3,10 @@ package rdx.works.profile.domain.account
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
+import rdx.works.profile.data.model.extensions.renameAccountDisplayName
 import rdx.works.profile.data.model.pernetwork.Network
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
-import rdx.works.profile.data.utils.renameAccountDisplayName
 import rdx.works.profile.di.coroutines.DefaultDispatcher
 import javax.inject.Inject
