+++ profile/src/main/java/rdx/works/profile/domain/depositguarantees/ChangeDefaultDepositGuaranteeUseCase.kt
@@ -3,9 +3,9 @@ package rdx.works.profile.domain.depositguarantees
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
+import rdx.works.profile.data.model.extensions.changeDefaultDepositGuarantee
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
-import rdx.works.profile.data.utils.changeDefaultDepositGuarantee
 import rdx.works.profile.di.coroutines.DefaultDispatcher
 import javax.inject.Inject
