+++ profile/src/main/java/rdx/works/profile/domain/gateway/AddGatewayUseCase.kt
@@ -2,7 +2,7 @@ package rdx.works.profile.domain.gateway
 
 import kotlinx.coroutines.flow.first
 import rdx.works.profile.data.model.apppreferences.Radix
-import rdx.works.profile.data.model.apppreferences.addGateway
+import rdx.works.profile.data.model.extensions.addGateway
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
 import javax.inject.Inject
