+++ profile/src/main/java/rdx/works/profile/domain/p2plink/DeleteP2PLinkUseCase.kt
@@ -1,7 +1,7 @@
 package rdx.works.profile.domain.p2plink
 
 import kotlinx.coroutines.flow.first
-import rdx.works.profile.data.model.apppreferences.deleteP2PLink
+import rdx.works.profile.data.model.extensions.deleteP2PLink
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
 import javax.inject.Inject
