+++ profile/src/main/java/rdx/works/profile/domain/signing/SignWithDeviceFactorSourceUseCase.kt
@@ -4,6 +4,7 @@ import com.radixdlt.ret.SignatureWithPublicKey
 import kotlinx.coroutines.flow.first
 import rdx.works.core.ret.toEngineModel
 import rdx.works.profile.data.model.deriveExtendedKey
+import rdx.works.profile.data.model.extensions.updateLastUsed
 import rdx.works.profile.data.model.factorsources.DeviceFactorSource
 import rdx.works.profile.data.model.pernetwork.Entity
 import rdx.works.profile.data.model.pernetwork.FactorInstance
@@ -12,7 +13,6 @@ import rdx.works.profile.data.model.pernetwork.SigningPurpose
 import rdx.works.profile.data.repository.MnemonicRepository
 import rdx.works.profile.data.repository.ProfileRepository
 import rdx.works.profile.data.repository.profile
-import rdx.works.profile.data.utils.updateLastUsed
 import rdx.works.profile.domain.NoMnemonicException
 import javax.inject.Inject
